requirements.txt
asgiref==3.2.3
autopep8==1.5
Babel==2.8.0
Django==3.0.4
django-environ==0.4.5
django-filter==2.2.0
django-formtools==2.2
django-phonenumber-field==4.0.0
django-tables2==2.3.1
djangorestframework==3.11.0
mysqlclient==1.4.6
phonenumbers==8.11.5
pycodestyle==2.5.0
pyquery==1.4.1
pytz==2019.3
sqlparse==0.3.1
